feat(k8s-kubeadm): Add support public ip which is not attached to an interface and open port 179/tcp for Calico BGP
Certain cloud providers provide public IP that is not attached to a
network interface. In this case, we have to use the address defined by
ansible_host in the inventory file. It is handled in 3 variables:
kubeadm_node_host, kubeadm_master_host, kubeadm_master_host_address.

- Remove ansible_ + interface ipv4 in all variables
- Maintain 2 variables for master host: kubeadm_master_host &
kubeadm_master_host_address
- Open 179/tcp port for Calico BGP